fix: Include eh_frame sections into the flash size
For ESP32-C3, C2, C6 chips, eh_frame sections are not placed into the
flash.rodata section but instead separately mapped to the
default_rodata_seg memory region.
This lead to the eh_frame sections not being calculated as part of the
flash_rodata size.

This commit fixes that with a simple workaround.