fix: enforce terminal control codes only if --output-file is not specified
Currently when "--force-terminal" is used along with "--output-file" it
will print terminal control codes/colors into the file. The purpose
of "--force-terminal" is to be used when esp-idf-size is started via
idf.py, so we can preserve the colors when the stdout is redirected.
It doesn't make much sense when the output is explicitly requested to
be placed in file with --output-file.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>