fix(ng): test that elf file exists before opening
Currently the elf file path is taken from project_description.json and
esp-idf-size fails if the file doesn't exist. This can happen if
e.g. linkage fails and the elf file is not available, but we can still
produce valid output based on the map file only. Just don't try
to read the elf file if it doesn't exist.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>