fix(ng): handle alignment/fill in section validation
Output section can have alignment/fill at the beginning, prior the
first input section. This can be seen e.g. in examples/wifi/iperf
for esp32.

.rtc.force_fast
                0x3ff80063        0x5
                0x3ff80064                        . = ALIGN (0x4)
 *fill*         0x3ff80063        0x1
                0x3ff80064                        _rtc_force_fast_start = ABSOLUTE (.)
                0x3ff80064                        _coredump_rtc_fast_start = ABSOLUTE (.)
 *(.rtc.fast.coredump .rtc.fast.coredump.*)
                0x3ff80064                        _coredump_rtc_fast_end = ABSOLUTE (.)
 *(.rtc.force_fast .rtc.force_fast.*)
 .rtc.force_fast.10
                0x3ff80064        0x1 esp-idf/esp_hw_support/libesp_hw_support.a(sleep_modes.c.obj)
                0x3ff80068                        . = ALIGN (0x4)
 *fill*         0x3ff80065        0x3
                0x3ff80068                        _rtc_force_fast_end = ABSOLUTE (.)

This fill is not accounted to any input section and since the validation
starts with output section address(0x3ff80063) it fails, because the
.rtc.force_fast.10 is expected at 0x3ff80063 and not at 0x3ff80064.

Fix this by using address of the first input section as starting address
for validation instead of the output section address.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>