fix(ng): provide more verbose error message when parsed map file validation fails
Currently the parsed map file is validated with _validate_sections().
If this fails esp-idf-size exits with a brief error e.g.

error:  input section .sdata is not at expected address 0x3fc94d99

This usually means that there is a bug in the map file parser. Let's
extend the error message and ask for a bug report and the map file for
further analyze.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>