fix(ng): correctly set input section stage in linker map parser
There are two cases how the input section can be entered in linker
map file.

1) all information at a single line
    .sdata         0x3fc94d94        0x4 /var/folders/3x/0ct1xt4n12jd90qfl3z4fp8m0000gn/T//ccMvH5l1.ltrans10.ltrans.o

2) information split into two lines
    .text.spi_flash_chip_th_get_caps
                   0x40394732        0x4 esp-idf/spi_flash/libspi_flash.a(spi_flash_chip_th.c.obj)

Which one is used depends on the input section name size. Linker
does the split on two lines if the name exceeds 15 chars.

Currently the parser is setting "in_input_section" state only
for the second case(facepalm). This may lead to a situation when a *fill*
following input section is not correctly accounted, because the parser
is not aware that it is in input section.

Fix this by setting the "in_input_section" state for both cases.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>