ci(ng): search for testing link map files recursively
The map files are already added to the esp-idf-size-test
repository with

https://github.com/espressif/esp-idf-size-test/commit/a3a00f67b880d38af4cd86dea82cce5c28332483

Since they are added into the legacy subdirectory, we need to make
sure they are picked up by the test, because it does not look
for map files recursively. Fix this by using rglob instead of glob.

This also adds rules in .gitlab-ci.yml to start ng test if there
is a change in chip_info yml files or ng tests.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>