fix: do not account noinit sections into total image size
Sections with the noinit suffix and .noinit do not contribute
to the total image size. These are NOLOAD/SHT_NOBITS sections which do
not undergo initialization during app startup and are not used during
application flashing. For example ".init", ".rtc_noinit" or ".ext_ram_noinit".
Let's skip these sections while counting the total image size.

For legacy version the total image size is counted based on already
filtered sections, so here we just make sure that the .noinit, which is
part of dram or diram, is not accounted.

New testing noinit_hello_world_esp32.map added. This adjusted hello_world
for esp32 with changes suggested in the 13515 issue.

Closes https://github.com/espressif/esp-idf/issues/13515

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>