fix(ng): fix memory map trimming for diff output
Currently the memory map is trimmed according to the command line
options for all output formats(tree, table, csv, ...). As a result
the diff output may be very confusing, because the trim function
also removes entries which were not changed. For example in diff
archive output the trimming removes archive sections which were
not changed. This is ok for the tree format, but for e.g. archive table
format this causes the total archive size to be reduced for the unchanged
sections.

For example

| Archive File  |   Total Size |  Flash Data |      .rodata |
| libmain.a     |     0   -292 |    0   -292 |     0   -292 |

The total size of libmain.a is wrong. This is because there was no
change in flash .text section between the two builds for the libmain.a
archive, only in .rodata, and the .text section for libmain.a is removed
by the trim function. Purpose of diff trimming is to show only changed
archives, but we cannot trim the memory map before the archive table
is created, because some sections may be removed. As a workaround
--show-unchanged could be used to mitigate this problem.

Let's remove the unchanged content based on the requested format.
Meaning remove unchanged archives only once the table is generated using
full memory map. With this approach the output is as expected.

| Archive File   |   Total Size | Flash Code |   .text | Flash Data |      .rodata |
| libmain.a      |     332 -292 | 332      0 |  332  0 |   0   -292 |     0   -292 |

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>