fix(ng): fix map and elf file detection
In early versions, the project_description.json file was used as an input,
so esp-idf-size can also use information from the project's ELF file.
Later, for backward compatibility and also because the ELF file may not be
produced if linkage fails, the input was switched to the map file and
the information from project_description.json and ELF, if available, is used
only as a complement to map file.

Currently when the map file is passed as input, esp-idf-size looks for
project_description.json in the same directory as the map file. If
found, it uses the "build_dir" when looking for the ELF file and also
for the map file. This causes problems when the project was
moved or renamed.

Fix this by looking for complement files only in the same directory
where the map file is.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>