fix: correctly handle DIRAM and overflows reports
As reported in the GH issue, the legacy version is wrongly
reporting overflow while running with booloader_override example.
The problem is how DIRAM total size is counted in the legacy version.
Currently the DIRAM total size available is divided by 2
if there is no IRAM available. This works correctly for specific cases, e.g.
where the whole IRAM is used for cache by setting the
CONFIG_ESP32S3_INSTRUCTION_CACHE_SIZE and if an alias memory region
exists. In the bootloader case, the IRAM is not used at all(not specified
in the linker script) and there is no memory range alias in DIRAM. Meaning the
legacy version is wrongly reducing(div by 2) the DIRAM available size,
which leads to the bogus overflow report. This is actually a similar problem
that 64ff3e19def0 ("idf_size.py: Fix issue where diram size was halved in..")
tried to fix.

There already was a fix to correctly account DIRAM
1f5d87857c37 ("idf_size.py: fixed diram counted twice issue, and imp..")
but it seems it missed one more step to work reliably. We cannot
blindly shrink DIRAM size without checking if two memory regions
actually overlap in DIRAM. There might be regions, as in the bootloader
example, that do not overlap in DIRAM and we should not shrink the DIRAM size,
because it leads to false overflow report. This commit also correctly splits
the memory regions based on the memory types defined in the yaml files.
Unfortunately this behaviour was, for a reason I don't know/understand,
reverted by b271033e2afc ("Tools: make idf_size work with overflow").
Note that the split si needed to correctly identify overlapping memory
regions.

This change brings the behavior from 1f5d87857c37 back, but adds a check
that two regions actually overlap before the DIRAM total size is reduced. It's
the same logic used by the refactored version. This should hopefully
allow the correct DIRAM and overflow handling in the legacy version.

Note that the output is changed because of the correct memory range
splitting based on the mermory types. But it seems this was already
presented in esp-idf-size for about a year between 1f5d87857c37 and
b271033e2afc, so maybe it's not a problem.

Here is the difference for hello_world on esp32s3

Output before change(actually same as generated by the refactored
version)

Total sizes:
Used static IRAM:   16383 bytes (      1 remain, 100.0% used)
      .text size:   15356 bytes
   .vectors size:    1027 bytes
Used stat D/IRAM:   57383 bytes ( 284377 remain, 16.8% used)
      .data size:   10968 bytes
      .bss  size:    2256 bytes
      .text size:   44159 bytes
Used Flash size :  161986 bytes
           .text:  116674 bytes
         .rodata:   45056 bytes
Total image size:  233496 bytes (.bin may be padded larger)

Output before this change

Total sizes:
Used static IRAM:   60542 bytes ( 297602 remain, 16.9% used)
      .text size:   59515 bytes
   .vectors size:    1027 bytes
Used stat D/IRAM:   13224 bytes ( 328536 remain, 3.9% used)
      .data size:   10968 bytes
      .bss  size:    2256 bytes
Used Flash size :  161986 bytes
           .text:  116674 bytes
         .rodata:   45056 bytes
Total image size:  233496 bytes (.bin may be padded larger)

In the old output the remaining DIRAM size is accounted twice, in
DIRAM and also in IRAM. In reality there isn't 297602 + 328536 bytes
available.

The change is related only to the summary, because detailed reports are
constructed differently. This can be seen in the diff for the expected
output. Also note that we are testing bootloader map files and we have
bogus overflows in the expected output, but this went unnoticed.
This is now fixed as can be seen in the updated expected output file
used for testing.

Closes https://github.com/espressif/esp-idf/issues/13727

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>