fix(ng): remove console width limit
The rich console is by default wrapping output to the console
width. This is making the output unusable in some situation.
For example when started via idf.py and RunTool as cmake target,
the console width on Windows detected by rich is reported as 80x25,
which is very small for e.g. archives report. The reason is that
the stdout is not attached to terminal(async RunTool), so the default
terminal size is used. Also the original condition setting the limit
was not correct if the stdout was redirected. This completely removes
the rich console wrapping and leaves it to terminal itself. This
behaves the same way as the legacy version. For more info
about the rich console wrapping please see
https://rich.readthedocs.io/en/stable/console.html#file-output

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>