fix(ng): fix line enumeration in link map file parser
Currently the line numbers used in error messages are
relative to the parsed link map section. For example
"Linker script and memory map". This is confusing and using
the line from the error message doesn't not point to the
correct place in the parsed map file. Fix this by using
global line numbers

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>