fix(ng): handle output sections with no content in map file
Address an issue where output sections with no content in the link map
file cause the parser to fail. These sections lack an address and size,
as seen in:

.xt.prop
  *(.xt.prop .xt.prop.* .gnu.linkonce.prop.*)

The parser currently assumes that if the address and size are not on the
same line as the output section name, they will be on the following
line. This assumption is incorrect for empty sections, leading to a
failure.

Additionally, note that the linker may split the output section
name/address/size line if the section name exceeds 16 characters.

This fix assigns a zero address and size to such sections.

Closes https://github.com/espressif/esp-idf/issues/13802

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>