change(ng): add documentation and --doc option
Currently, options and API documentation are available solely via
argparse and pydoc. This update introduces markdown documentation,
offering more comprehensive details about the refactored version.

Given that the documentation is in markdown and the refactored version
utilizes rich, which can render markdown, a new --doc option is added to
display it.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>