fix: handle archive paths that contain spaces
The archive path in the input section might include spaces. At present,
the tool divides the input section information based on whitespace,
which can result in unexpected entries after splitting. To address this,
use the maxsplit argument to limit the split, as the archive path and
object name are the final entries in the input section record.

Closes https://github.com/espressif/esp-idf-size/issues/5

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>