feat: add archives' dependencies report
Introduce a new reporting option, --archives-dependencies, which
enables the display of archive dependencies or dependents. It relies on
the "Cross Reference Table" found in the link map file, included by the
ESP-IDF build system through the use of the -cref linker option. The
format is straightforward, as specified in the ld documentation:

    For  each symbol, a list of file names is given.  If the symbol is
    defined, the  first  file  listed  is  the location  of  the definition.
    If the symbol is defined as a common value then any files where this
    happens appear next.   Finally  any files that reference the symbol are
    listed.

<symbol> <definition archive>(<definition object file>)
         <reference archive>(<reference object file>)
         <reference archive>(<reference object file>)
         ...

With this information, it's possible to identify dependency
relationships between archives. Unfortunately, the cross-reference table
also includes discarded symbols. This issue is addressed by filtering
symbols based on the symbol table in the ELF file and by removing
archives that are not used in the "Linker script and memory map" section
of the link map file.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>