ci(ng): fix artifacts generation in test_memmap.py
The esp-idf-size tool no longer uses project_description.json as input;
instead, it now uses the map file, and project_description.json path is
derived from this map file. Although this change has been implemented
for a while, it appears that new artifacts have not been generated since
then. To address this, ensure that the link map file is passed to
esp-idf-size when generating artifacts.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>