feat: adjust esp32h4 chip info description
During development, the original esp32h4 was renamed, but its chip_info
yaml was retained to avoid disrupting the existing esp-idf-size
implementation. Now that the esp32h4 name is being reused, the existing
chip_info description needs to be updated to reflect the current memory
layout of the esp32h4. This update also eliminates the exclusion of the
esp32h4 chip from testing and includes the addition of relevant test
artifacts.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>