fix: add a note about total memory type sizes reported in summary
To minimize potential questions about the total sizes mentioned in the
summary report, include a note clarifying that these values are based on
the memory region sizes specified in the linker script. They may not
correspond to the sizes outlined in the TRM, as some portions of memory
may be reserved. Also, the total flash size does not reflect other flash
usage and application partition size.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>