fix(ng): exclude the total flash size from the report by default
The total flash size available for the application is currently
determined by the flash memory region size specified in the map file.
This size is set by the linker script to the maximum possible flash
size, which often does not match the actual flash size defined by
`CONFIG_ESPTOOLPY_FLASHSIZE`. The reported total flash size does not
accurately represent the flash space available for the application
because the flash also contains other data, such as the bootloader and
partition table. Additionally, the application's flash size is
constrained by its partition size. As a result, it is not possible to
reliably determine the total flash size available for the application.
Exclude the total flash size from the report to prevent any potential
confusion about its meaning.

Closes https://github.com/espressif/esp-idf/issues/15915

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>